# [TTS] FastPitch training: speed up align_prior_matrix calculation