--- a/requirements/requirements_common.txt
+++ b/requirements/requirements_common.txt
@@ -2,3 +2,4 @@ sentencepiece<1.0.0
 youtokentome>=1.0.5
 pandas
 sacremoses>=0.0.43
+einops