--- a/requirements/requirements_torch_tts.txt
+++ b/requirements/requirements_torch_tts.txt
@@ -5,3 +5,4 @@ pystoi
 pesq
 pandas
 inflect
+einops