# What does this PR do ?

Speed up align_prior_matrix calculation and don't save it to disk. They take lots of space (tens of gigabytes per experiment) and it is perfectly fine to compute them on-the-fly. Matrix calculation speed-up is significant.

```python
%timeit beta_binomial_prior_distribution_numpy_old(80, 300)
226 ms ± 5.86 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit beta_binomial_prior_distribution_new_torch(80, 300)
586 µs ± 22 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

**Collection**: TTS

# Changelog 
- Add specific line by line info of high level changes in this PR.

# Usage
* You can potentially add a usage example below



# Before your PR is "Ready for review"
**Pre checks**:
- [ ] Make sure you read and followed [Contributor guidelines](https://github.com/NVIDIA/NeMo/blob/main/CONTRIBUTING.md)
- [ ] Did you write any new necessary tests?
- [ ] Did you add or update any necessary documentation?
- [ ] Does the PR affect components that are optional to install? (Ex: Numba, Pynini, Apex etc)
  - [ ] Reviewer: Does the PR have correct import guards for all optional libraries?
  
**PR Type**:
- [ ] New Feature
- [ ] Bugfix
- [ ] Documentation

If you haven't finished some of the above items you can still open "Draft" PR.


## Who can review?

Anyone in the community is free to review the PR once the checks have passed. 
[Contributor guidelines](https://github.com/NVIDIA/NeMo/blob/main/CONTRIBUTING.md) contains specific people who can review PRs to various areas.

# Additional Information
* Related to # (issue)