@racoiaws 2022-08-10
I'd also remove the `BetaBinomialInterpolator` class, apparently it exists just for caching `beta_binomial_prior_distribution`