@racoiaws 2022-08-25
I did a closer-to-real benchmark: iterate through LibriTTS train-clean-100 dataset (one epoch), `batch size=32`, `num_workers=16`, no training

Baseline: 
- 800s 1st epoch, 77s per epoch onwards, +14GB for a 9G dataset
- 800s per epoch when not saving to SSD

BetaBinomialInterpolator: 
- 114s per epoch

PyTorch:
- 72s per epoch, same with no mel spec calculation



So, PyTorch implementation is much faster than baseline; on par with caching but no extra space and I/O. Also, `self.get_log_mel(audio).shape[2]` is good enough, no need to replace it with explicit formula

Will be updating the PR shortly